add changelog entry
